// DashboardApp/src/ib-plugins/table/report/query/types.ts
export type SortedParamsType = {
  unsatisfiedByParentParams: ParameterItem[]
  satisfiedByParentParams: JSONRecord
}